Fix merging of process.env in Node 4+
`builder envs` did not work on Node 4+ because `_.merge` was behaving
differently due to `process.env` no longer being a plain object. This
clones `shOpts` (which contains `env`) so that merge works as expected.

I considered `_.toPlainObject(process.env)` in `environment.js` in place
of this fix, to fix the problem at its source. That worked, but I was
fearful of the implications as Environment explicitly says that `env` is
"Environment object to mutate" - and using that fix we'd no longer be
mutating `process.env` but a copy.

Fixes #63

/cc @ryan-roemer